Reorganize React components
