+++ apps/spotlight/src/pages/form-sample.astro
@@ -1,7 +1,7 @@
 ---
 import { createSequentialInterview } from '@atj/interviews';
 import { getAppContext } from '../context';
-import { InterviewForm } from '../components/react/interview-form';
+import { InterviewForm } from '../components/react/experiments/interview-form';
 import ContentLayout from '../layouts/ContentLayout.astro';
 
 const context = getAppContext(import.meta.env);
