+++ apps/spotlight/src/pages/import-document.astro
@@ -1,6 +1,5 @@
 ---
-import { createSequentialInterview } from '@atj/interviews';
-import { DocumentImporter } from '../components/react/document-importer';
+import { DocumentImporter } from '../components/react/form-builder/document-importer';
 import { getAppContext } from '../context';
 import ContentLayout from '../layouts/ContentLayout.astro';
