+++ apps/spotlight/src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import { DocumentAssembler } from '../components/react/document-assembler';
+import { DocumentAssembler } from '../components/react/experiments/document-assembler';
 import { getAppContext } from '../context';
 import ContentLayout from '../layouts/ContentLayout.astro';
