+++ apps/spotlight/src/pages/ud105-evicition-form.astro
@@ -1,6 +1,6 @@
 ---
 import { getAppContext } from '../context';
-import DynamicForm from '../components/react/dynamic-form';
+import DynamicForm from '../components/react/form';
 import formData from '../htmlParser/ud105-form-field-output.json';
 import ContentLayout from '../layouts/ContentLayout.astro';
